Notes page (#2)
Completed this pull request on Tarun's behalf as it has been approved by Dr. Chen and I need to clone the project to my machine. 

* Created Notebook file

Currently empty

* First version

Has a main nav menu, a list group to navigate Notebooks, and a div that will hold a flexbox where all the notes will be shown.

* Put image in separate folder

* Moved image to separate folder